Break out start, scoutgroupinfo, persons
